% dataWorkshop
% Copyright 2004-2013, Institute for Mathematics Applied Geosciences
% University Corporation for Atmospheric Research
% Licensed under the GPL -- www.gpl.org/licenses/gpl.html

\name{dataWorkshop}
\alias{dataWorkshop-package}
\alias{dataWorkshop}
\title{
dataWorkshop  
}
\description{
This a package that collects some substantial data sets
 to use for teaching data analysis in R. It also loads the 
 \code{fields} package that has some useful functions for  working with 
 spatial  data.
\describe{
A listing of the names and type of data
\item{\link{BoulderDaily}}{Daily temperatures and rainfall for Boulder 1987- 2014.}
\item{\link{COmonthlyMet}}{Monthly temperature and precipitation from 1895-1997 for several hundred stations in Colorado.}
\item{\link{CO2}}{Global CO2 concentrations simulated by a global model.} 
\item{\link{DC3}}{Chemical species measured along the flight paths of two research aircraft}
\item{\link{Denver0Z}}{Radiosonde temperature measurements for the
                        Denver upper air station made at "0Z GMT"
                        Greenwhich mean time ( early evening in
                        Denver).}
\item{\link{narccap1}}{Seasonal Temperatures from the North American Regional Climate Change and Assessment Program (NARCCAP) Phase I }
\item{\link{narccap2}}{Seasonal Temperatures from NARCCAP Phase II}
\item{\link{NorthAmericanRainfall}}{Average summer total rainfall for approx. 1700 North American stations}
\item{\link{PilotSUV}}{Asking prices of approximately 200 used Honda Pilot SUVs with model, year and mileage}
\item{\link{PRISMelevation}}{Elevations for the NorthAmerica on a 4km grid}
\item{\link{RMelevation}}{Elevations for the Rocky Mountain Region on a 4km grid}
\item{\link{SST}}{Monthly and gridded sea surface temperatures for the
                        Pacific Ocean}
\item{\link{UStornado}}{Reported date,  locations and severity of tornados in the US 1950-2012}
\item{\link{VictoriaT}}{Space-time  monthly images of Lake Victoria temperatures}
\item{\link{WorldBankCO2}}{CO2 Emissions per capita with country level demographics on population and per captia income}
\item{\link{YSUsers}}{User activity on the Yellowstone supercomputer}
}

}
\keyword{datasets}
