% fields, Tools for spatial data
% Copyright 2004-2013, Institute for Mathematics Applied Geosciences
% University Corporation for Atmospheric Research
% Licensed under the GPL -- www.gpl.org/licenses/gpl.html
\name{miniTriathlon}
\alias{miniTriathlon}
\title{
  Mini triathlon results
}
\description{
Results from a mini-triathlon sponsored by Bud Lite, held in
Cary, NC, June 1990. Times are in minutes for the male 30-34 
group. Man was it hot and humid! The events in  order were:
swim:  (1/2 mile)
bike:  (15 miles)
run:   (4 miles)

Data set is a 110X3 matrix (note row names are not important)
with columns (1:3) indicating the 3 events (swim, bike, run) 
and rows indexing participants. 

}
\examples{
data(miniTriathlon)
 pairs( miniTriathlon)
}

\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
